Add signal handling to agbot/css_start.sh
When run in a container anax is run via a shell script that does not
handle SIGTERM and hence when the container is stopped this signal does
not arrive to the binary for graceful exit. A consequence of this is
when run via docker compose up, docker compose down will have to wait
until the timeout in order to kill the process.

The same happens for the ccs-api.

This change adds proper SIGTERM handling to agbot_start.sh and
css_start.sh.

Signed-off-by: Leonardo Bergesio <lbergesio@users.noreply.github.com>